SPL/requirements.txt
# Core dependencies for SpecuLab SPL project
numpy>=1.20.0
astropy>=5.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
scipy>=1.9.0
scikit-image>=0.19.0
PyYAML>=5.4.0

# Optional GPU acceleration (uncomment if you have CUDA-capable GPU)
# cupy-cuda11x>=11.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x

# Note: specula package needs to be installed separately
# This is a custom package that may need to be installed from source
# or from a specific repository

